A variety of fixes in bean contexts removal of slf4j-jdk14 artifact from solr projects adding staging location mappings to fedora.fcfg file (space separated mappings, separated by pipe) having trouble with logging from fedora modules..

+++ persistence/src/main/java/edu/unc/lib/dl/services/DigitalObjectManagerImpl.java
@@ -155,8 +155,10 @@ public abstract class DigitalObjectManagerImpl implements DigitalObjectManager {
 			return result;
 		} catch (IngestException e) {
 			// exception on AIP preparation, no transaction started
-			log.info("User level exception on ingest, prior to Fedora transaction", e);
-			aip.delete();
+			log.error("Exception during submission or enqueuing", e);
+			if(aip != null) {
+				aip.delete();
+			}
 			throw e;
 		}
 	}
